requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data and APIs
yfinance>=0.2.18
alpha_vantage>=2.3.1
quandl>=3.7.0
pandas-datareader>=0.10.0
requests>=2.31.0
websocket-client>=1.6.0

# Technical Analysis
TA-Lib>=0.4.25
pandas-ta>=0.3.14b0
talib-binary>=0.4.19

# Machine Learning & Deep Learning
xgboost>=1.7.0
catboost>=1.2.0
lightgbm>=4.0.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
optuna>=3.2.0

# Time Series
statsmodels>=0.14.0
arch>=5.5.0
pykalman>=0.9.5

# Risk Management
pyfolio>=0.9.2
empyrical>=0.5.5
quantlib>=1.31

# Logging and Configuration
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Numerical computing
numba>=0.57.0
cython>=3.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0